requirements.txt
boto
pypng
numpy
pyfits
